[Render] rename files to avoid name collisions in the IDE
