+++ Library/Lyra/Render.hpp
@@ -2,11 +2,11 @@
 #define LYRA_LIBRARY_RENDER_HPP
 
 // RHI (Render Hardware Interface)
-#include <Lyra/Render/RHI/API.h>
-#include <Lyra/Render/RHI/Enums.h>
-#include <Lyra/Render/RHI/Descs.h>
-#include <Lyra/Render/RHI/Types.h>
-#include <Lyra/Render/RHI/Inits.h>
+#include <Lyra/Render/RHI/RHIAPI.h>
+#include <Lyra/Render/RHI/RHIEnums.h>
+#include <Lyra/Render/RHI/RHIDescs.h>
+#include <Lyra/Render/RHI/RHITypes.h>
+#include <Lyra/Render/RHI/RHIInits.h>
 
 // RPI (Render Pass Interface)
 #include <Lyra/Render/RPI/FrameGraph.h>
@@ -17,9 +17,9 @@
 #include <Lyra/Render/RPI/FrameGraphResource.h>
 
 // SLC (Shader Language Compiler)
-#include <Lyra/Render/SLC/API.h>
-#include <Lyra/Render/SLC/Enums.h>
-#include <Lyra/Render/SLC/Descs.h>
-#include <Lyra/Render/SLC/Types.h>
+#include <Lyra/Render/SLC/SLCAPI.h>
+#include <Lyra/Render/SLC/SLCEnums.h>
+#include <Lyra/Render/SLC/SLCDescs.h>
+#include <Lyra/Render/SLC/SLCTypes.h>
 
 #endif // LYRA_LIBRARY_RENDER_HPP
