+++ Library/Lyra/Render/RPI/FrameGraphBuilder.h
@@ -8,7 +8,7 @@
 #include <Lyra/Render/RPI/FrameGraphEnums.h>
 #include <Lyra/Render/RPI/FrameGraphResource.h>
 
-namespace lyra::rpi
+namespace lyra::rhi
 {
     struct FrameGraph;
 
@@ -105,6 +105,6 @@ namespace lyra::rpi
         uint            pass  = 0xFFFFFFFFu;
     }; // end of FrameGraphBuilder
 
-} // namespace lyra::rpi
+} // namespace lyra::rhi
 
 #endif // LYRA_LIBRARY_RENDER_RPI_FRAME_GRAPH_BUILDER_H
