+++ Library/Lyra/Render/SLC/SLCTypes.h
@@ -7,11 +7,11 @@
 #include <Lyra/Common/String.h>
 #include <Lyra/Common/Handle.h>
 #include <Lyra/Common/Pointer.h>
-#include <Lyra/Render/RHI/Enums.h>
-#include <Lyra/Render/RHI/Descs.h>
-#include <Lyra/Render/SLC/Enums.h>
-#include <Lyra/Render/SLC/Utils.h>
-#include <Lyra/Render/SLC/Descs.h>
+#include <Lyra/Render/RHI/RHIEnums.h>
+#include <Lyra/Render/RHI/RHIDescs.h>
+#include <Lyra/Render/SLC/SLCEnums.h>
+#include <Lyra/Render/SLC/SLCUtils.h>
+#include <Lyra/Render/SLC/SLCDescs.h>
 
 namespace lyra::rhi
 {
