+++ Plugins/D3D12/D3D12Utils.h
@@ -15,8 +15,8 @@ using Microsoft::WRL::ComPtr;
 #include <Lyra/Common/Slotmap.h>
 #include <Lyra/Common/Container.h>
 #include <Lyra/Common/Compatibility.h>
-#include <Lyra/Render/RHI/Descs.h>
-#include <Lyra/Render/RHI/API.h>
+#include <Lyra/Render/RHI/RHIDescs.h>
+#include <Lyra/Render/RHI/RHIAPI.h>
 
 using namespace lyra;
 using namespace lyra::rhi;
@@ -867,7 +867,7 @@ inline void ThrowIfFailed(HRESULT hr)
         auto err = get_hresult_message(hr);
         if (err.c_str())
             get_logger()->error("error: {}\n", err.c_str());
-        throw std::exception(); 
+        throw std::exception();
     }
 }
