+++ Plugins/D3D12/Plugin.cpp
@@ -3,7 +3,7 @@
 // global module headers
 #include <Lyra/Common/String.h>
 #include <Lyra/Common/Plugin.h>
-#include <Lyra/Render/RHI/API.h>
+#include <Lyra/Render/RHI/RHIAPI.h>
 #include <Lyra/Window/WSIAPI.h>
 
 using namespace lyra;
