+++ Plugins/Glfw/Plugin.cpp
@@ -22,7 +22,7 @@
 #include <Lyra/Common/Plugin.h>
 #include <Lyra/Window/WSIAPI.h>
 #include <Lyra/Window/WSITypes.h>
-#include <Lyra/Render/RHI/Types.h>
+#include <Lyra/Render/RHI/RHITypes.h>
 
 using namespace lyra;
 using namespace lyra::rhi;
