+++ Plugins/Vulkan/Plugin.cpp
@@ -1,7 +1,7 @@
 // global module headers
 #include <Lyra/Common/String.h>
 #include <Lyra/Common/Plugin.h>
-#include <Lyra/Render/RHI/API.h>
+#include <Lyra/Render/RHI/RHIAPI.h>
 #include <Lyra/Window/WSIAPI.h>
 
 #include "VkUtils.h"
